feat: cfd-422 hide delete option behind a new feature flag (#174)
* hide delete option behind a new feature flag

* add a snapshot test

* redirect to error on delete api when delete feature is disabled

* rename variable to a better name and fix logic

* fix snapshot test

* use redirect to error function

+++ repos/fdbt-site/src/pages/api/deletePurchaseMethod.ts
@@ -2,20 +2,30 @@ import { NextApiResponse } from 'next';
 import { deleteSalesOfferPackageByNocCodeAndName } from '../../data/auroradb';
 import { redirectToError, redirectTo, getAndValidateNoc } from '../../utils/apiUtils/index';
 import { NextApiRequestWithSession } from '../../interfaces';
+import { globalSettingsDeleteEnabled } from '../../../src/constants/featureFlag';
 
 export default async (req: NextApiRequestWithSession, res: NextApiResponse): Promise<void> => {
-    try {
-        const { query } = req;
-        const id = Number(query?.id);
+    if (globalSettingsDeleteEnabled) {
+        try {
+            const { query } = req;
+            const id = Number(query?.id);
 
-        if (!Number.isInteger(id)) {
-            throw new Error('insufficient data provided for delete query');
+            if (!Number.isInteger(id)) {
+                throw new Error('insufficient data provided for delete query');
+            }
+            const nocCode = getAndValidateNoc(req, res);
+            await deleteSalesOfferPackageByNocCodeAndName(id, nocCode);
+
+            redirectTo(res, '/viewPurchaseMethods');
+        } catch (error) {
+            const message = 'There was a problem deleting the selected SOP:';
+
+            redirectToError(res, message, 'api.deletePurchaseMethod', error);
         }
-        const nocCode = getAndValidateNoc(req, res);
-        await deleteSalesOfferPackageByNocCodeAndName(id, nocCode);
-        redirectTo(res, '/viewPurchaseMethods');
-    } catch (error) {
-        const message = 'There was a problem deleting the selected SOP:';
-        redirectToError(res, message, 'api.deletePurchaseMethod', error);
+    } else {
+        redirectToError(res, 'Cannot perform the delete operation', 'api.deletePurchaseMethod', {
+            name: 'delete',
+            message: 'cannot delete',
+        });
     }
 };
